Use INT in examples instead of i64.

+++ examples/hello.rs
@@ -1,9 +1,9 @@
-use rhai::{Engine, EvalAltResult};
+use rhai::{Engine, EvalAltResult, INT};
 
 fn main() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    let result = engine.eval::<i64>("40 + 2")?;
+    let result = engine.eval::<INT>("40 + 2")?;
 
     println!("Answer: {}", result); // prints 42
